pyproject.toml
[project]
name = "rag-mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "chromadb>=1.3.5",
    "docx2txt>=0.9",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-google-genai>=2.0.10",
    "mcp>=1.22.0",
    "numpy>=2.3.5",
    "pathlib>=1.0.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
]
